requirements.txt
Flask
pymongo
authlib
requests
python-dotenv
